requirements.txt
Scrapy==1.1.0rc3
pymongo==3.2.1
PyYAML
